fix:
AddOrEditListing component:
Allow map coordinates to be populated optionally.

Listing component:
Display the listing's map coordinates.

SearchListings component:
The price ranges collection must be created as an array of objects, each with price range and index as properties.

LocationsProvider hook:
Record up to 10 previous url locations.

firestoreComms.js:
Move the function transformListingData to this file, so as to have one centralised instance, instead of same function existing across many components.

Reports component:
Import the function transformListingData that was moved to firestoreComms.js file so as to have one reusable copy.

Listings component: 
Determine correctly when to re-load listings when the user is exploring listings. So that the user is served with the correct listings per main place explored.
Import the function transformListingData that was moved to firestoreComms.js file so as to have one reusable copy.